% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_time_cutoff.R
\name{apply_time_cutoff}
\alias{apply_time_cutoff}
\title{Apply a Time Cutoff to Survival Data}
\usage{
apply_time_cutoff(
  data,
  target_info,
  id = NULL,
  time_cutoff = NULL,
  traceon = FALSE
)
}
\arguments{
\item{data}{A data frame containing the survival data.}

\item{target_info}{A named vector with elements \code{id}, \code{time}, \code{event}, and \code{type}:
\itemize{
\item \code{id} : string identifying target
\item \code{time}: Name of the time column in \code{data}.
\item \code{event}: Name of the event column in \code{data}.
\item \code{type}: Censoring type, either \code{"right"} or \code{"mstate"}.
}}

\item{id}{Optional. Column name in the data, specifying the individual identifiers. Default is \code{NULL}.}

\item{time_cutoff}{Optional. Numeric value specifying the time cutoff at which to truncate the time variable. If \code{NULL}, returns the original data.}

\item{traceon}{Logical. If \code{TRUE}, enables tracing for debugging purposes. Default is \code{FALSE}.}
}
\value{
A data.table with survival times truncated at the specified \code{time_cutoff}.
}
\description{
The \code{apply_time_cutoff} function truncates survival data at a specified time cutoff and returns the modified dataset.
}
\note{
This function employs \code{survival::survSplit()} and is useful for preparing survival data for analyses requiring a time cutoff.
}
\examples{
# Load the required packages
library(survival)
library(data.table)
data(cancer, package="survival")
lung$status_01 <- lung$status -1
lung$status <- NULL
head(lung)

# Define the target_info list
target_info <- c(time = "time", event = "status_01", type = "right")

# Apply the time cutoff
lungx <- apply_time_cutoff(lung, target_info, time_cutoff = 365)

# View the first few rows of the transformed data
head(lungx)

}
