% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_example.R
\name{run_example}
\alias{run_example}
\title{Run and List Example Scripts}
\usage{
run_example(script_prefix = NULL, purl = FALSE)
}
\arguments{
\item{script_prefix}{A character string representing the beginning of the
script's name. If NULL, the function lists all available scripts.}

\item{purl}{A logical value indicating whether to generate an R script from the
R Markdown file. If TRUE, a script is created with a "-purl" postfix. Defaults to FALSE.}
}
\value{
If no prefix is given, returns a character vector of available script
names. If a valid prefix is provided and the script is rendered, opens the
HTML content in a web browser and returns the path to the HTML file.
}
\description{
This function lists available example scripts or renders a specified script
as HTML using the \code{spin} function from the knitr package. It also optionally
generates an R script from the rendered R Markdown file using the \code{purl} function.
The HTML content, if generated, is automatically opened in the default web browser.
}
\examples{
# List all available example scripts
run_example()

# Run and render an example script as HTML and open it in the default web browser
# Assuming a script named "01*.R" exists
\dontrun{
run_example("01")

# Run and render an example script and also generate a purled R script
run_example("01", purl = TRUE)
}
}
