% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceit.R
\name{traceit}
\alias{traceit}
\title{Trace Function for Debugging}
\usage{
traceit(msg, variable = NULL)
}
\arguments{
\item{msg}{A character string representing the message to be printed.}

\item{variable}{An optional variable whose structure is printed to help with debugging; defaults to \code{NULL}.}
}
\description{
This internal function prints trace messages to the console if tracing is enabled. It is designed to
help with debugging by outputting specified messages and, optionally, the structure of a given variable.
}
\details{
The function is controlled by the \code{traceon} flag, which is set elsewhere in the code. When \code{traceon} is \code{TRUE},
the function uses \code{cat} to print the provided message. If a variable is also provided, it prints the variable
using \code{print()}. This is particularly useful for tracing control flow and inspecting variable states during development.
}
\keyword{internal}
